extracting interface
